+++ draft-packages/guidance-block/docs/GuidanceBlock.stories.tsx
@@ -20,12 +20,7 @@ export default {
       "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=1929%3A39077"
     ),
   },
-  decorators: [
-    withDesign,
-    story => (
-      <div style={{ display: "flex", justifyContent: "center" }}>{story()}</div>
-    ),
-  ],
+  decorators: [withDesign, story => <div>{story()}</div>],
 }
 
 const guidanceBlockText = {
@@ -50,7 +45,6 @@ const Default = () => (
         onClick: () => alert("tada: 🎉"),
       },
     }}
-    noMaxWidth
   />
 )
 
@@ -85,6 +79,18 @@ const WithoutMaxWidth = () => (
     illustration={<Informative alt="" />}
     text={guidanceBlockText}
     noMaxWidth
+    actions={{
+      primary: {
+        label: "Action",
+        onClick: () => {
+          alert("tada: 🎉")
+        },
+      },
+      secondary: {
+        label: "Secondary action",
+        href: "#",
+      },
+    }}
   />
 )
 
@@ -166,7 +172,6 @@ const WithCustomDescription = () => (
       },
     }}
     persistent
-    noMaxWidth
   />
 )