+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.9.4...@kaizen/draft-guidance-block@2.9.5) (2021-05-20)
+
+
+### Bug Fixes
+
+* Make GuidanceBlock action buttons right aligned ([#1621](https://github.com/cultureamp/kaizen-design-system/issues/1621)) ([0dfebb2](https://github.com/cultureamp/kaizen-design-system/commit/0dfebb245ebd1cb6a0484758437d92d7061e5936))
+
+
+
+
+
+## [2.9.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.9.3...@kaizen/draft-guidance-block@2.9.4) (2021-05-19)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.9.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.9.2...@kaizen/draft-guidance-block@2.9.3) (2021-05-18)
+
+
+### Bug Fixes
+
+* guidance block max-width not centering description text ([#1610](https://github.com/cultureamp/kaizen-design-system/issues/1610)) ([c80185b](https://github.com/cultureamp/kaizen-design-system/commit/c80185b4eb89197820514709f3ac6d211f8141aa))
+
+
+
+
+
 ## [2.9.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.9.1...@kaizen/draft-guidance-block@2.9.2) (2021-05-18)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block